Cleaned up prefabs and scripts and organized them
