--- 380Guantlet/Assets/Scripts/Blaster.cs
@@ -1,31 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Blaster : MonoBehaviour
-{
-    private void Update()
-    {
-        BlasterActive();
-    }
-
-    public void BlasterActive()
-    {
-        transform.Translate(new Vector3(0, -5 * Time.deltaTime, 0));
-    }
-
-    private void OnCollisionEnter(Collision collision)
-    {
-        if (collision.gameObject.tag == "Enemy")
-        {
-            Destroy(collision.gameObject);
-            StartCoroutine(ResetBlaster());
-            Destroy(this.gameObject);
-        }
-    }
-
-    IEnumerator ResetBlaster()
-    {
-        yield return new WaitForSeconds(2);
-    }
-}
